+++ src/main/java/it/ids/hackathown/api/dto/request/AddMentorRequest.java
@@ -3,5 +3,5 @@ package it.ids.hackathown.api.dto.request;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
-public record AddMentorRequest(@NotNull @Min(1) Long mentorUserId) {
+public record AddMentorRequest(@NotNull @Min(1) Integer mentorUserId) {
 }
