+++ src/main/java/it/ids/hackathown/api/dto/request/RegisterTeamRequest.java
@@ -3,5 +3,5 @@ package it.ids.hackathown.api.dto.request;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
 
-public record RegisterTeamRequest(@NotNull @Min(1) Long teamId) {
+public record RegisterTeamRequest(@NotNull @Min(1) Integer teamId) {
 }
