+++ src/main/java/it/ids/hackathown/api/dto/response/EvaluationResponse.java
@@ -4,10 +4,10 @@ import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record EvaluationResponse(
-    Long id,
-    Long hackathonId,
-    Long submissionId,
-    Long judgeId,
+    Integer id,
+    Integer hackathonId,
+    Integer submissionId,
+    Integer judgeId,
     BigDecimal score0to10,
     String comment,
     LocalDateTime createdAt
