+++ src/main/java/it/ids/hackathown/api/dto/response/SubmissionResponse.java
@@ -3,8 +3,8 @@ package it.ids.hackathown.api.dto.response;
 import java.util.Date;
 
 public record SubmissionResponse(
-    Long id,
-    Long registrationId,
+    Integer id,
+    Integer registrationId,
     String title,
     String description,
     String repoUrl,
