+++ src/main/java/it/ids/hackathown/api/dto/response/SupportRequestResponse.java
@@ -4,9 +4,9 @@ import it.ids.hackathown.domain.enums.StatoRichiesta;
 import java.time.LocalDateTime;
 
 public record SupportRequestResponse(
-    Long id,
-    Long hackathonId,
-    Long teamId,
+    Integer id,
+    Integer hackathonId,
+    Integer teamId,
     String message,
     LocalDateTime createdAt,
     StatoRichiesta status
