+++ src/main/java/it/ids/hackathown/api/dto/response/WinnerResponse.java
@@ -3,8 +3,8 @@ package it.ids.hackathown.api.dto.response;
 import java.time.LocalDateTime;
 
 public record WinnerResponse(
-    Long hackathonId,
-    Long teamId,
+    Integer hackathonId,
+    Integer teamId,
     LocalDateTime declaredAt,
     String paymentTxId
 ) {
